2021/ctf-10-22-2021/web-jwt/requirements.txt
Flask==1.1.2
pyjwt==1.4.2